Move TCP types into net::tcp module
